Split `CreatePlatform` handler into a dedicated `create.go` file for better separation of concerns.
