+++ api/platform/create.go
@@ -1,0 +1,45 @@
+package platformHandler
+
+import (
+	"encoding/json"
+	"net/http"
+	"products/internal/db"
+	"time"
+
+	"github.com/jackc/pgx/v5/pgtype"
+)
+
+type CreatePlatformRequest struct {
+	Name        string `json:"name"`
+	Description string `json:"description"`
+}
+
+func (h *PlatformHandler) CreatePlatform(w http.ResponseWriter, r *http.Request) {
+	var req CreatePlatformRequest
+	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
+		http.Error(w, "Invalid request body", http.StatusBadRequest)
+		return
+	}
+
+	if req.Name == "" {
+		http.Error(w, "Name is required", http.StatusBadRequest)
+		return
+	}
+
+	if err := h.queries.CreatePlatform(r.Context(), db.CreatePlatformParams{
+		Name: req.Name,
+		Description: pgtype.Text{
+			Valid:  req.Description != "",
+			String: req.Description,
+		},
+		Timestamp: pgtype.Timestamptz{
+			Valid: true,
+			Time:  time.Now().UTC(),
+		},
+	}); err != nil {
+		http.Error(w, "Failed to create platform", http.StatusInternalServerError)
+		return
+	}
+
+	w.WriteHeader(http.StatusCreated)
+}
